requirements.txt
Django==1.8
djangorestframework==3.2
psycopg2==2.6.1